+++ main.js
@@ -1,19 +1,27 @@
+/*global brackets, define, $ */
 define(function (require, exports, module) {
 	'use strict';
 
-	var Commands         = brackets.getModule('command/Commands');
-	var CommandManager   = brackets.getModule('command/CommandManager');
-	var Menus            = brackets.getModule('command/Menus');
-	var DocumentManager  = brackets.getModule('document/DocumentManager');
-	var FileSystem       = brackets.getModule('filesystem/FileSystem');
-	var ExtensionUtils   = brackets.getModule('utils/ExtensionUtils');
-	var WorkspaceManager = brackets.getModule('view/WorkspaceManager');
-	var DefaultDialogs   = brackets.getModule('widgets/DefaultDialogs');
-	var Dialogs          = brackets.getModule('widgets/Dialogs');
-	var UglifyJS         = require('UglifyJS/uglifyjs');
-	window.MOZ_SourceMap = require('./SourceMap/source-map');
+	var Commands           = brackets.getModule('command/Commands'),
+		CommandManager     = brackets.getModule('command/CommandManager'),
+		Menus              = brackets.getModule('command/Menus'),
+		DocumentManager    = brackets.getModule('document/DocumentManager'),
+		FileSystem         = brackets.getModule('filesystem/FileSystem'),
+		PreferencesManager = brackets.getModule('preferences/PreferencesManager'),
+		ProjectManager     = brackets.getModule('project/ProjectManager'),
+		ExtensionUtils     = brackets.getModule('utils/ExtensionUtils'),
+		WorkspaceManager   = brackets.getModule('view/WorkspaceManager'),
+		DefaultDialogs     = brackets.getModule('widgets/DefaultDialogs'),
+		Dialogs            = brackets.getModule('widgets/Dialogs'),
+		UglifyJS           = require('UglifyJS/uglifyjs'),
+		
+		menu = Menus.getMenu(Menus.AppMenuBar.FILE_MENU),
+		bottomPanel = null,
+		panelLog = null,
+		toolbarIcon = null,
+		pendingTasks = 0;
 	
-	var pendingTasks = 0;
+	window.MOZ_SourceMap   = require('./SourceMap/source-map');
 
 	// Log
 	function log(s) {
@@ -29,7 +37,10 @@ define(function (require, exports, module) {
 		pendingTasks -= 1;
 		if (pendingTasks < 1) {
 			appendLog('Done!<br/>');
-			toolbarIcon.removeAttribute('class');
+			if (toolbarIcon.getAttribute('class') === 'active') {
+				toolbarIcon.setAttribute('class', 'success');
+			}
+			window.setTimeout(function () {toolbarIcon.removeAttribute('class'); }, 3000);
 		}
 	}
 
@@ -38,15 +49,17 @@ define(function (require, exports, module) {
 		bottomPanel.show();
 		
 		// Get current directory
-		var directory = DocumentManager.getCurrentDocument().file.parentPath;
-		var path = directory + output;
+		var directory = DocumentManager.getCurrentDocument().file.parentPath,
+			path = directory + output,
 		
 		// Start UglifyJS magic!
-		var ast = null;
-		var compressor = UglifyJS.Compressor();
-		var source_map = UglifyJS.SourceMap();
-		var stream = UglifyJS.OutputStream({source_map: source_map});
-		var i, l;
+			ast = null,
+			code = null,
+			compressor = UglifyJS.Compressor(),
+			source_map = UglifyJS.SourceMap(),
+			stream = UglifyJS.OutputStream({source_map: source_map}),
+			i = 0,
+			l = 0;
 		for (i = 0, l = inputs.length; i < l; i += 1) {
 			appendLog('Parsing file: ' + inputs[i].name);
 			ast = UglifyJS.parse(inputs[i].content, {filename: inputs[i].name, toplevel: ast});
@@ -62,7 +75,7 @@ define(function (require, exports, module) {
 		}
 		appendLog('Extracting...');
 		ast.print(stream);
-		var code = stream.toString();
+		code = stream.toString();
 		
 		// Append isolation code
 		if (options && options.isolate) {
@@ -74,14 +87,14 @@ define(function (require, exports, module) {
 		// Save the map
 		if (!options || options.generateMap) {
 			code += '\n//# sourceMappingURL=' + output.split('/').pop() + '.map';
-			var map = source_map.toString();
 			pendingTasks += 1;
-			FileSystem.getFileForPath(path + '.map').write(map, {blind: true}, function (err) {
+			FileSystem.getFileForPath(path + '.map').write(source_map.toString(), {blind: true}, function (err) {
 				if (err) {
 					appendLog('Error generating map at:<br/>' + path + '.map');
-					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'JS Compiler', 'Error on map genetarion:\n' + err);
+					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_ERROR, 'JS Compiler', 'Error on map genetarion:\n' + err);
+					toolbarIcon.setAttribute('class', 'error');
 				} else {
-					appendLog('Map successfully generated at:<br/>\n' + path + '.map');
+					appendLog('Map successfully generated at:<br/>' + path + '.map');
 				}
 				taskDone();
 			});
@@ -92,7 +105,8 @@ define(function (require, exports, module) {
 		FileSystem.getFileForPath(path).write(code, {blind: true}, function (err) {
 			if (err) {
 				appendLog('Error compiling code at:<br/>' + path);
-				Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'JS Compiler', 'Error on compilation:\n' + err);
+				Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_ERROR, 'JS Compiler', 'Error on compilation:\n' + err);
+				toolbarIcon.setAttribute('class', 'error');
 			} else {
 				appendLog('File successfully compiled at:<br/>' + path);
 			}
@@ -107,7 +121,8 @@ define(function (require, exports, module) {
 			
 			FileSystem.getFileForPath(directory + options.inputs[counter]).read({}, function (err, content) {
 				if (err) {
-					appendLog('Error on reading this file: ' + err);
+					appendLog('Error on reading ' + options.inputs[counter] + ': ' + err + '<br/>' + options.output + ' compilation cancelled.<br/>');
+					toolbarIcon.setAttribute('class', 'warning');
 				} else {
 					// Read content
 					contents.push({name: options.inputs[counter], content: content});
@@ -118,10 +133,40 @@ define(function (require, exports, module) {
 			// Contents where collected
 			if (contents.length > 0) {
 				// Finally compile the code
-				doUglify(contents, options.output, options);
+				if (options.precompile) {
+					appendLog('Precompiling code into: precompiled.js');
+					var content = '',
+						i = 0,
+						l = 0;
+					for (i = 0, l = contents.length; i < l; i += 1) {
+						content += contents[i].content + '\n';
+					}
+					doUglify([{name: 'precompiled.js', content: content}], options.output, options);
+				} else {
+					doUglify(contents, options.output, options);
+				}
 			} else {
 				appendLog('Something went wrong.<br/>Done!');
 			}
+		}
+	}
+	
+	function compileWithOptions(content, directory) {
+		var options = JSON.parse(content),
+			i = 0,
+			l = 0;
+		if (options.outputs) {
+			// Compile each output in options
+			l = options.outputs.length;
+			appendLog('Found ' + l + ' outputs');
+			for (i = 0; i < l; i += 1) {
+				appendLog('Generating ' + options.outputs[i].output);
+				getContentsFrom(options.outputs[i], directory, [], 0);
+			}
+		} else {
+			// Compile with old single output option format
+			appendLog('Generating ' + options.output);
+			getContentsFrom(options, directory, [], 0);
 		}
 	}
 	
@@ -131,43 +176,39 @@ define(function (require, exports, module) {
 		toolbarIcon.setAttribute('class', 'active');
 		
 		// Search for options file
-		var currentFile = DocumentManager.getCurrentDocument().file;
-		var directory = currentFile.parentPath;
+		var currentFile = DocumentManager.getCurrentDocument().file,
+			directory = currentFile.parentPath,
+			preferences = null;
 		FileSystem.getFileForPath(directory + '.jscompiler.json').read({}, function (err, content) {
 			if (err) {
 				if (err === 'NotFound') {
-					// Options file not found. Try to compile current file
+					// Options file not found. Using project options?
 					bottomPanel.show();
-					appendLog('No options file. Compiling current script');
-					var ext = currentFile.name.split('.').pop();
-					if (ext === 'js') {
-						doUglify([{name: currentFile.name, content: DocumentManager.getCurrentDocument().getText()}], currentFile.name.replace(/\.js$/, '.min.js'));
+					preferences = PreferencesManager.get('jscompiler');
+					if (preferences !== undefined) {
+						// Read project options file
+						appendLog('Loading project options');
+						compileWithOptions(preferences, directory);
 					} else {
-						// Current file is not JavaScript. Warn!
-						appendLog('Current document is not JavaScript');
+						// Options not found. Try to compile current file
+						appendLog('No options file. Compiling current script');
+						var ext = currentFile.name.split('.').pop();
+						if (ext === 'js') {
+							doUglify([{name: currentFile.name, content: DocumentManager.getCurrentDocument().getText()}], currentFile.name.replace(/\.js$/, '.min.js'));
+						} else {
+							// Current file is not JavaScript. Warn!
+							appendLog('Current document is not JavaScript');
+						}
 					}
 				} else {
-					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'JS Compiler', 'Error on reading options:\n' + err);
+					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_ERROR, 'JS Compiler', 'Error on reading options:\n' + err);
+					toolbarIcon.setAttribute('class', 'error');
 				}
 			} else {
-				// Open options template
+				// Read portable options file
 				bottomPanel.show();
-				appendLog('Loading options');
-				var options = JSON.parse(content);
-				if (options.outputs) {
-					// Compile each output in options
-					var i, l;
-					l = options.outputs.length;
-					appendLog('Found ' + l + ' outputs');
-					for (i = 0; i < l; i += 1) {
-						appendLog('Generating ' + options.outputs[i].output);
-						getContentsFrom(options.outputs[i], directory, [], 0);
-					}
-				} else {
-					// Compile with old single output option format
-					appendLog('Generating ' + options.output);
-					getContentsFrom(options, directory, [], 0);
-				}
+				appendLog('Loading portable options');
+				compileWithOptions(content, directory);
 			}
 		});
 	}
@@ -176,44 +217,80 @@ define(function (require, exports, module) {
 		// Read the template
 		ExtensionUtils.loadFile(module, 'templates/.jscompiler.json').then(function (result) {
 			// Get directory and file name for project options
-			var file = DocumentManager.getCurrentDocument().file;
-			var directory = file.parentPath;
-			var directories = directory.split('/');
-			directories.pop();
-			var dirname = directories.pop();
-			var filename = file.name.replace(/\.js$/, '');
+			var code = null,
+				jsonCode = null,
+				file = DocumentManager.getCurrentDocument().file,
+				directory = file.parentPath,
+				directories = directory.split('/'),
+				dirname = directories.pop(),
+				filename = file.name.replace(/\.js$/, ''),
+				i = 0,
+				j = 0,
+				l = 0,
+				jl = 0;
+			window.console.log(directory);
 			if (filename === file.name) {
 				filename = 'script';
 			}
-			var code = result.replace(/%DIRECTORY%/g, dirname).replace(/%FILENAME%/g, filename);
-			// Create options file
-			FileSystem.getFileForPath(directory + '.jscompiler.json').write(code, {blind: true}, function (err) {
-				if (err) {
-					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'JS Compiler', 'Error on creating options:\n' + err);
-				} else {
-					// Open options file
-					CommandManager.execute(Commands.CMD_OPEN, {fullPath: directory + '.jscompiler.json'});
+			dirname = directories.pop();
+			code = result.replace(/%DIRECTORY%/g, dirname).replace(/%FILENAME%/g, filename);
+			
+			// Ask the user which class of options file wants to use
+			Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'Options file not found', 'Select the type of options file you want to use for this project:', [{className: Dialogs.DIALOG_BTN_CLASS_LEFT, id: Dialogs.DIALOG_BTN_CANCEL, text: 'Cancel'}, {className: Dialogs.DIALOG_BTN_CLASS_PRIMARY, id: 'project', text: 'Project'}, {className: Dialogs.DIALOG_BTN_CLASS_NORMAL, id: 'portable', text: 'Portable'}]).done(function (id) {
+				window.console.log('ID: ' + id);
+				if (id === 'project') {
+					// Create options in project file
+					jsonCode = JSON.parse(code);
+					directory = ProjectManager.makeProjectRelativeIfPossible(directory);
+					for (i = 0, l = jsonCode.outputs.length; i < l; i += 1) {
+						// Assign relative path to each input and output
+						for (j = 0, jl = jsonCode.outputs[i].inputs.length; j < jl; j += 1) {
+							jsonCode.outputs[i].inputs[j] = directory + jsonCode.outputs[i].inputs[j];
+						}
+						jsonCode.outputs[i].output = directory + jsonCode.outputs[i].output;
+					}
+					PreferencesManager.set('jscompiler', jsonCode, {location: {scope: 'project'}});
+					// Open project options file
+					window.setTimeout(function () {CommandManager.execute(Commands.CMD_OPEN, {fullPath: ProjectManager.getProjectRoot().fullPath + '.brackets.json'}); }, 200);
+				} else if (id === 'portable') {
+					// Create portable options file
+					FileSystem.getFileForPath(directory + '.jscompiler.json').write(code, {blind: true}, function (err) {
+						if (err) {
+							Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_ERROR, 'JS Compiler', 'Error on creating options:\n' + err);
+						} else {
+							// Open options file
+							CommandManager.execute(Commands.CMD_OPEN, {fullPath: directory + '.jscompiler.json'});
+						}
+					});
 				}
 			});
 		});
 	}
 	
 	function openOptions() {
-		var directory = DocumentManager.getCurrentDocument().file.parentPath;
-		var path = directory + '.jscompiler.json';
+		var directory = DocumentManager.getCurrentDocument().file.parentPath,
+			path = directory + '.jscompiler.json',
+			preferences = null;
 		
 		// Reading the options file content
 		FileSystem.getFileForPath(path).read({}, function (err, content) {
 			if (err) {
 				if (err === 'NotFound') {
-					// Options file not found. Generate
-					log('Options file not found. Creating!');
-					generateOptions();
+					// Options file not found. Using project options?
+					preferences = PreferencesManager.get('jscompiler');
+					if (preferences !== undefined) {
+						// Open project options file
+						CommandManager.execute(Commands.CMD_OPEN, {fullPath: ProjectManager.getProjectRoot().fullPath + '.brackets.json'});
+					} else {
+						// Options not found. Generate
+						log('Options file not found. Creating!');
+						generateOptions();
+					}
 				} else {
-					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'JS Compiler', 'Error on reading options:\n' + err);
+					Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_ERROR, 'JS Compiler', 'Error on reading options:\n' + err);
 				}
 			} else {
-				// Open options file
+				// Open portable options file
 				CommandManager.execute(Commands.CMD_OPEN, {fullPath: path});
 			}
 		});
@@ -225,7 +302,7 @@ define(function (require, exports, module) {
 		// Reading the options template... Just to be sure no one has deleted it accidentally...
 		FileSystem.getFileForPath(path).read({}, function (err, content) {
 			if (err) {
-				Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_INFO, 'JS Compiler', 'Error on reading template:\n' + err);
+				Dialogs.showModalDialog(DefaultDialogs.DIALOG_ID_ERROR, 'JS Compiler', 'Error on reading template:\n' + err);
 			} else {
 				// Open options template
 				CommandManager.execute(Commands.CMD_OPEN, {fullPath: path});
@@ -238,7 +315,6 @@ define(function (require, exports, module) {
 	}
 	
 	// Add file menu option
-	var menu = Menus.getMenu(Menus.AppMenuBar.FILE_MENU);
 	menu.addMenuDivider();
 	
 	// Register commands
@@ -252,15 +328,15 @@ define(function (require, exports, module) {
 	menu.addMenuItem('jscompiler.template');
 	
 	// Start bottom panel
-	var bottomPanel = WorkspaceManager.createBottomPanel('jscompiler.panel', $("<div id='jscompiler-panel' class='bottom-panel vert-resizable top-resizer' style='box-sizing: border-box; height: 200px; display: block;'><div class='toolbar simple-toolbar-layout'><div class='title'>JSCompiler</div> <a href='#' class='close'>×</a></div><div id='log' class='table-container resizable-content' style='height: 170px'></div></div></div>"));
-	var panelLog = bottomPanel.$panel.find('#log');
+	bottomPanel = WorkspaceManager.createBottomPanel('jscompiler.panel', $('<div id="jscompiler-panel" class="bottom-panel vert-resizable top-resizer" style="box-sizing: border-box; height: 200px; display: block;"><div class="toolbar simple-toolbar-layout"><div class="title">JSCompiler</div> <a href="#" class="close">×</a></div><div id="log" class="table-container resizable-content" style="height: 170px"></div></div></div>'));
+	panelLog = bottomPanel.$panel.find('#log');
 	bottomPanel.$panel.find('.close').on('click', closePanel);
 	
 	// Load css
 	ExtensionUtils.loadStyleSheet(module, 'styles/main.css');
 	
 	// Add toolbar icon
-	var toolbarIcon = $('<a>')
+	toolbarIcon = $('<a>')
 		.attr({
 			id: 'toolbar-jscompiler',
 			title: 'Compress JavaScript',